quote_message/requirements.txt
openai>=1.0.0
schedule
python-dotenv
